Enhance query get external purchase order

+++ Com.DanLiris.Service.Purchasing.Lib/Facades/VBRequestPOExternal/IVBRequestPOExternalService.cs
@@ -8,7 +8,7 @@ namespace Com.DanLiris.Service.Purchasing.Lib.Facades.VBRequestPOExternal
     public interface IVBRequestPOExternalService
     {
         List<POExternalDto> ReadPOExternal(string keyword, string division, string currencyCode);
-        List<SPBDto> ReadSPB(string keyword, string division, List<int> epoIds, string currencyCode);
+        List<SPBDto> ReadSPB(string keyword, string division, List<long> epoIds, string currencyCode);
         int UpdateSPB(string division, int spbId);
         Task<int> AutoJournalVBRequest(VBFormDto form);
     }
